refactor: move cubit in bloc
